+++ CHANGES.rst
@@ -5,6 +5,7 @@ Released 2024-10-21
 
 -   Support Python 3.13
 -   Deprecate Python 3.8
+-   Add limit option to unique text.
 
 Version 0.1.1
 -------------